+++ src/components/takes/options/TakeExportMenuItem.jsx
@@ -5,15 +5,13 @@ import AddToDriveIcon from '@material-symbols/svg-400/rounded/add_to_drive.svg';
 import DownloadIcon from '@material-symbols/svg-400/rounded/download.svg';
 
 import { useGAPITokenHandler } from '@/lib/googleapi';
-import { getVideoBlob } from '@/recorder/cache/VideoCache';
+import { useCachedVideoBlob } from '@/recorder/cache';
 import {
   useTakeDownloader,
   useTakeGoogleDriveUploader,
 } from '@/serdes/UseTakeExporter';
 import { useTakeExportedIDBKey } from '@/stores/DocumentStoreContext';
 import MenuStyle from '@/styles/Menu.module.css';
-
-import { useForceRefreshOnMenuOpen } from './MenuItemHelper';
 
 /**
  * @param {object} props
@@ -28,10 +26,8 @@ export default function TakeCacheMenuItem({
   shotId,
   takeId,
 }) {
-  // NOTE: It is important that cache checks are done sparingly (only when needed)
-  const menuOpen = useForceRefreshOnMenuOpen();
+  const videoBlob = useCachedVideoBlob(documentId, takeId);
 
-  const [cached, setCached] = useState(false);
   const [googleDriveConnected, setGoogleDriveConnected] = useState(false);
   const idbKey = useTakeExportedIDBKey(documentId, takeId);
   const downloadTake = useTakeDownloader();
@@ -39,47 +35,39 @@ export default function TakeCacheMenuItem({
   const handleToken = useGAPITokenHandler();
 
   function onDownloadClick() {
-    if (!idbKey) {
+    if (!idbKey || !videoBlob) {
       return;
     }
     downloadTake(documentId, sceneId, shotId, takeId);
   }
 
   function onGoogleDriveClick() {
+    if (!videoBlob) {
+      return;
+    }
     uploadTake(documentId, sceneId, shotId, takeId);
   }
 
   useEffect(() => {
-    // Test video in cache
-    getVideoBlob(takeId)
-      .then((result) => {
-        if (result) {
-          setCached(true);
-        } else {
-          setCached(false);
-        }
-      })
-      .catch(() => setCached(false));
-
     // Test google drive connection
     if (!handleToken((token) => setGoogleDriveConnected(Boolean(token)))) {
       setGoogleDriveConnected(false);
     }
-  }, [menuOpen, handleToken, takeId]);
+  }, [handleToken, takeId]);
 
   return (
     <div className="flex flex-row">
       <MenuItem
         className={MenuStyle.menuItem + ' ' + 'flex-1 flex flex-row'}
         onClick={onDownloadClick}
-        disabled={!cached}>
+        disabled={!videoBlob}>
         <span className="flex-1">Export to</span>
         <DownloadIcon className="w-6 h-6 fill-current" />
       </MenuItem>
       <MenuItem
         className={MenuStyle.menuItem}
         onClick={onGoogleDriveClick}
-        disabled={!cached || !googleDriveConnected}>
+        disabled={!videoBlob || !googleDriveConnected}>
         <AddToDriveIcon className="w-6 h-6 fill-current" />
       </MenuItem>
     </div>