+++ src/recorder/cache/VideoCacheContext.jsx
@@ -1,0 +1,5 @@
+import { createContext } from '@/utils/ReactContextHelper';
+
+import { useVideoCacheContextValue } from './VideoCacheContextValue';
+
+export const VideoCacheContext = createContext(useVideoCacheContextValue);