+++ src/utils/StringFormat.js
@@ -18,3 +18,19 @@ export function formatBytes(bytes, minimum = 0) {
     BYTE_UNITS[BYTE_UNITS.length - 1]
   }`;
 }
+
+/**
+ * @param {number} millis
+ */
+export function formatHourMinSecTime(millis) {
+  const seconds = Math.floor((millis / 1000) % 60);
+  const minutes = Math.floor((millis / (1000 * 60)) % 60);
+  const hours = Math.floor((millis / (1000 * 60 * 60)) % 24);
+  return (
+    String(hours).padStart(2, '0') +
+    ':' +
+    String(minutes).padStart(2, '0') +
+    ':' +
+    String(seconds).padStart(2, '0')
+  );
+}